pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "lipsync-automation"
version = "2.0.0"
description = "Psycho-cinematic automation system for lip-sync video generation"
readme = "README.md"
authors = [{name = "Ishan Lagesh", email = "ishan@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "Pillow",
    "moviepy",
    "librosa",
    "soundfile",
    "onnxruntime",
    "python-dotenv",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "isort",
    "pre-commit",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
]

[project.scripts]
lipsync = "lipsync_automation.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lipsync_automation*"]

[tool.setuptools.package-data]
lipsync_automation = ["config/*.json", "assets/**/*"]